# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "药房进销存与处方调配后端服务",
        "title": "药房管理系统 API",
        "contact": {},
        "version": "1.3.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/special-drugs/ampoule-returns": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "special-drugs"
                ],
                "summary": "空安瓿回收记录",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "药品ID",
                        "name": "drug_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "special-drugs"
                ],
                "summary": "空安瓿回收登记",
                "parameters": [
                    {
                        "description": "回收记录",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/yaofang_internal_model.AmpouleReturn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            }
        },
        "/special-drugs/ampoule-returns/{id}/verify": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "special-drugs"
                ],
                "summary": "核对空安瓿回收",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "回收记录ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            }
        },
        "/special-drugs/dispense-register": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "special-drugs"
                ],
                "summary": "发药专册登记（补录）",
                "parameters": [
                    {
                        "description": "专账记录",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/yaofang_internal_model.SpecialDrugLedger"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            }
        },
        "/special-drugs/ledgers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "special-drugs"
                ],
                "summary": "专账查询",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "药品ID",
                        "name": "drug_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "记录类型",
                        "name": "log_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "开始时间",
                        "name": "start",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "结束时间",
                        "name": "end",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            }
        },
        "/special-drugs/prescriptions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "special-drugs"
                ],
                "summary": "麻精处方登记列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "患者姓名",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页条数",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Body"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "internal_handler.Body": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "yaofang_internal_model.AmpouleReturn": {
            "type": "object",
            "properties": {
                "batch_no": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "drug_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "patient_name": {
                    "type": "string"
                },
                "prescription_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "return_date": {
                    "type": "string"
                },
                "returned_by": {
                    "type": "string"
                },
                "status": {
                    "description": "pending/verified",
                    "type": "string"
                },
                "verified_by": {
                    "type": "string"
                }
            }
        },
...
        "yaofang_internal_model.SpecialDrugLedger": {
            "type": "object",
            "properties": {
                "batch_no": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "drug_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "log_type": {
                    "description": "prescribe/dispense/return/ampoule_return/usage",
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "operator_id": {
                    "type": "integer"
                },
                "operator_name": {
                    "type": "string"
                },
                "patient_card_no": {
                    "type": "string"
                },
                "patient_name": {
                    "type": "string"
                },
                "prescription_id": {
                    "type": "integer"
                },
                "quantity": {
                    "description": "有符号",
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}